Refactoring: Move ObservationsInspector out of Map component
