+++ app/containers/ExplorePage.tsx
@@ -18,6 +18,7 @@ import ObservationsTable from '../components/ObservationsTable';
 import ExplorerFilter from '../components/explorerFilters';
 import ExplorerMetrics from '../components/explorerMetrics';
 import { EmptyClasses, RareAnimalsClasses } from '../constants/animalsClasses';
+import ObservationsInspector from '../components/ObservationsInspector';
 
 type Filters = {
   activeAnimals: Entry[];
@@ -90,6 +91,7 @@ export default function ExplorePage() {
     certaintyRange: [0, 1]
   });
   const [data, setData] = useState<undefined | ObservationsData>();
+  const [inspectedObservations, setInspectedObservations] = useState<Observation[]>([]);
   const [predictionOverrides, setPredictionOverrides] = useState<Record<string, string>>({});
 
   const handleFilters = (val: string[]) => {
@@ -146,11 +148,21 @@ export default function ExplorePage() {
       >
         <Card style={{ height: '100%' }} interactive elevation={Elevation.TWO}>
           <Callout intent={Intent.PRIMARY}>{t('explore.mapHint')}</Callout>
-          <Map
-            data={getFilteredData(data)}
-            predictionOverrides={predictionOverrides}
-            onPredictionOverride={handlePredictionOverride}
-          />
+          <div
+            style={{
+              width: '100%',
+              position: 'relative',
+              overflow: 'hidden'
+            }}
+          >
+            <Map data={getFilteredData(data)} onInspect={setInspectedObservations}/>
+            <ObservationsInspector
+              observations={inspectedObservations}
+              onClose={() => setInspectedObservations([])}
+              predictionOverrides={predictionOverrides}
+              onPredictionOverride={handlePredictionOverride}
+            />
+          </div>
         </Card>
       </div>
     </div>
